+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 1.1.0 - 27 May 2022
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `RuleBuilderExtensions.IsDateTime` method.
+- Added `RuleBuilderExtensions.IsDateTimeOrEmpty` method.
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.0.1 - 26 July 2021
 
 Breaking changes: